+++ consensus/core/src/lib.rs
@@ -24,8 +24,6 @@ pub mod trusted;
 pub mod tx;
 pub mod utxo;
 
-pub mod testutils;
-
 /// Integer type for accumulated PoW of blue blocks. We expect no more than
 /// 2^128 work in a single block (btc has ~2^80), and no more than 2^64
 /// overall blocks, so 2^192 is definitely a justified upper-bound.
