--- consensus/core/src/testutils/mod.rs
@@ -1,2 +1,0 @@
-pub mod create_transaction;
-pub mod op_true_script;
