+++ crypto/addresses/Cargo.toml
@@ -10,6 +10,7 @@ license.workspace = true
 borsh.workspace = true
 serde.workspace = true
 smallvec.workspace = true
+thiserror.workspace = true
 
 [dev-dependencies]
 criterion.workspace = true
