+++ crypto/txscript/src/standard.rs
@@ -74,6 +74,55 @@ pub fn extract_script_pub_key_address(script_public_key: &ScriptPublicKey, prefi
     }
 }
 
+pub mod test_helpers {
+    use super::*;
+    use crate::{
+        opcodes::codes::{OpPushData1, OpTrue},
+        MAX_TX_IN_SEQUENCE_NUM,
+    };
+    use consensus_core::{
+        constants::TX_VERSION,
+        subnets::SUBNETWORK_ID_NATIVE,
+        tx::{Transaction, TransactionInput, TransactionOutpoint, TransactionOutput},
+    };
+
+    // op_true_script returns a P2SH script paying to an anyone-can-spend address,
+    // The second return value is a redeemScript to be used with txscript.PayToScriptHashSignatureScript
+    pub fn op_true_script() -> (ScriptPublicKey, Vec<u8>) {
+        let redeem_script = vec![OpTrue];
+        let script_public_key = pay_to_script_hash_script(&redeem_script);
+        (script_public_key, redeem_script)
+    }
+
+    // create_transaction create a transaction that spends the first output of provided transaction.
+    // Assumes that the output being spent has opTrueScript as it's scriptPublicKey.
+    // Creates the value of the spent output minus provided `fee` (in sompi).
+    pub fn create_transaction(tx_to_spend: &Transaction, fee: u64) -> Transaction {
+        let (script_public_key, redeem_script) = op_true_script();
+
+        // TODO: call txscript.PayToScriptHashSignatureScript(redeemScript, nil) when available
+        let signature_script = pay_to_script_hash_signature_script(redeem_script, vec![]);
+
+        let previous_outpoint = TransactionOutpoint::new(tx_to_spend.id(), 0);
+        let input = TransactionInput::new(previous_outpoint, signature_script, MAX_TX_IN_SEQUENCE_NUM, 1);
+        let output = TransactionOutput::new(tx_to_spend.outputs[0].value - fee, script_public_key);
+        Transaction::new(TX_VERSION, vec![input], vec![output], 0, SUBNETWORK_ID_NATIVE, 0, vec![])
+    }
+
+    pub fn pay_to_script_hash_signature_script(redeem_script: Vec<u8>, signature: Vec<u8>) -> Vec<u8> {
+        // This is just a hack until we have ScriptBuilder
+        // For now, we just suppose that redeem_script len is in the u8 range
+        Vec::from_iter(
+            signature
+                .iter()
+                .copied()
+                .chain(once(OpPushData1))
+                .chain(once(redeem_script.len() as u8))
+                .chain(redeem_script.iter().copied()),
+        )
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
